+++ enterprise/admin/forms.py
@@ -30,6 +30,7 @@ from enterprise.api_client.discovery import CourseCatalogApiClient
 from enterprise.api_client.lms import EnrollmentApiClient
 from enterprise.models import (
     EnterpriseCustomer,
+    EnterpriseCustomerCatalog,
     EnterpriseCustomerIdentityProvider,
     EnterpriseCustomerReportingConfiguration,
 )
@@ -319,6 +320,7 @@ class EnterpriseCustomerAdminForm(forms.ModelForm):
             "slug",
             "country",
             "active",
+            "customer_type",
             "site",
             "catalog",
             "enable_data_sharing_consent",
@@ -487,6 +489,11 @@ class EnterpriseCustomerReportingConfigAdminForm(forms.ModelForm):
     This form uses the PasswordInput widget to obscure passwords as they are
     being entered by the user.
     """
+    enterprise_customer_catalogs = forms.ModelMultipleChoiceField(
+        EnterpriseCustomerCatalog.objects.all(),
+        required=False,
+    )
+
     class Meta:
         model = EnterpriseCustomerReportingConfiguration
         fields = (
@@ -507,11 +514,36 @@ class EnterpriseCustomerReportingConfigAdminForm(forms.ModelForm):
             "sftp_username",
             "decrypted_sftp_password",
             "sftp_file_path",
+            "enterprise_customer_catalogs",
         )
         widgets = {
             'decrypted_password': forms.widgets.PasswordInput(),
             'decrypted_sftp_password': forms.widgets.PasswordInput(),
         }
+
+    def clean(self):
+        """
+        Override of clean method to perform additional validation
+        """
+        cleaned_data = super(EnterpriseCustomerReportingConfigAdminForm, self).clean()
+        report_customer = cleaned_data.get('enterprise_customer')
+
+        # Check that any selected catalogs are tied to the selected enterprise.
+        invalid_catalogs = [
+            '{} ({})'.format(catalog.title, catalog.uuid)
+            for catalog in cleaned_data.get('enterprise_customer_catalogs')
+            if catalog.enterprise_customer != report_customer
+        ]
+
+        if invalid_catalogs:
+            message = _(
+                'These catalogs for reporting do not match enterprise'
+                'customer {enterprise_customer}: {invalid_catalogs}',
+            ).format(
+                enterprise_customer=report_customer,
+                invalid_catalogs=invalid_catalogs,
+            )
+            self.add_error('enterprise_customer_catalogs', message)
 
 
 class TransmitEnterpriseCoursesForm(forms.Form):